Rename netx -> listener
